Remove game directory, move Facing and ScriptRef to engine
Game-specific states (title, gameplay, pause, transition) and systems
(player movement, NPC patrol, etc.) are removed. Games will be built
as separate projects using the engine as a dependency.

Facing and ScriptRef components moved to default_components.hpp
since the scripting engine depends on them.
